service/ingestion-service/requirements.txt
paho-mqtt
redis
minio
sqlalchemy
psycopg2-binary
python-dotenv
fastapi
uvicorn[standard]